change fileSize field type from Int to Long in Animation, Audio, Document, File, Video, Voice types.

+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/Audio.kt
@@ -8,6 +8,6 @@ data class Audio(
     val title: String? = null,
     val fileName: String? = null,
     val mimeType: String? = null,
-    val fileSize: Int? = null,
+    val fileSize: Long? = null,
     val thumb: PhotoSize?,
 )
