+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/File.kt
@@ -3,6 +3,6 @@ package eu.vendeli.tgbot.types
 data class File(
     val fileId: String,
     val fileUniqueId: String,
-    val fileSize: Int? = null,
+    val fileSize: Long? = null,
     val filePath: String? = null,
 )
